refactor eth_sender unittests
