+++ core/lib/zksync_core/src/eth_sender/mod.rs
@@ -10,9 +10,6 @@ mod zksync_functions;
 #[cfg(test)]
 mod tests;
 
-#[cfg(test)]
-mod tests_helpers;
-
 pub use self::{
     aggregator::Aggregator, error::ETHSenderError, eth_tx_aggregator::EthTxAggregator,
     eth_tx_manager::EthTxManager,
